requirements.txt
discord.py==2.2.2
openai==0.27.4
python-dotenv==1.0.0
azure.cognitiveservices.speech==1.28.0
